Code cleanup and comments; add name to attachment thumbnail in Preview mode; remove macCatalyst restrictions for QuickLook and iOS 16 restrictions for photo picker; rename and group files in project folders.
