+++ Sources/ArcGISToolkit/Common/ThumbnailView.swift
@@ -12,8 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import SwiftUI
 import ArcGIS
+import SwiftUI
 
 /// A view displaying a thumbnail image for an attachment.
 struct ThumbnailView: View  {
