+++ Sources/ArcGISToolkit/Extensions/ArcGIS/AttachmentHelpers/PopupAttachment.swift
@@ -16,6 +16,7 @@ import ArcGIS
 import Foundation
 
 extension PopupAttachment : FeatureAttachment {
+    /// The type of the attachment.
     public var featureAttachmentKind: FeatureAttachmentKind {
         FeatureAttachmentKind(kind: kind)
     }
